Committing version 1.0.1
Fixed an error where an update to the JavaScript engine would cause an
XML validation error. This was caused by JavaScript-characteristics.js,
CPITracker.groovy and updates.xsd not agreeing on which characteristic
ID to use for the JavaScript engine. This has now been aligned.

Please note that updates.xsd has been modified, so if you use this file
anywhere, please update accordingly.